refactored to build pypi distribution
(cherry picked from commit 6d32b4a2a0adfb6ec9e61ba179e551d11b436134)
